Initial text for usrguide3
Most of the material has come from xparse.dtx, but
re-ordered and somewhat re-worded to aim at the
target audience. The remaining concepts will either
go into l3cmd.dtx and thus to interface3, or will be
retained in xparse.dtx (for the deprectated ideas).

More examples for users are likely needed in usrguide3:
these could be how to implement current document
commands, for example.